requirements.txt
django===2.2
djangorestframework==3.10.3
django-cors-headers==3.4.0
python-json-logger==0.1.11
mysqlclient
uwsgi
requests